auth_service/tests/functional/requirements-dev.txt
pytest-flask==1.2.0